+++ android/src/main/java/ly/count/dart/countly_flutter/CountlyFlutterPlugin.java
@@ -1303,6 +1303,18 @@ public class CountlyFlutterPlugin implements MethodCallHandler, FlutterPlugin, A
 
                 Countly.sharedInstance().views().stopAllViews(segmentation);
                 result.success(null);
+            } else if ("addSegmentationToViewWithID".equals(call.method)) {
+                String viewId = args.getString(0);
+                Map<String, Object> segmentation = toMap(args.getJSONObject(1));
+
+                // Countly.sharedInstance().views().addSegmentationToViewWithID(viewId, segmentation);
+                result.success(null);
+            } else if ("addSegmentationToViewWithName".equals(call.method)) {
+                String viewName = args.getString(0);
+                Map<String, Object> segmentation = toMap(args.getJSONObject(1));
+
+                // Countly.sharedInstance().views().addSegmentationToViewWithName(viewName, segmentation);
+                result.success(null);
             } else if ("startAutoStoppedView".equals(call.method)) {
                 String viewName = args.getString(0);
                 Map<String, Object> segmentation = toMap(args.getJSONObject(1));